Removed HttpStoreSwapper and related tests, as it is dead code.
